apps/data-service/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.23.0
pydantic>=2.2.0
pytest>=8.5.0
pytest-asyncio>=0.23.0
httpx>=0.25.0
loguru>=0.7.0
beautifulsoup4>=4.12.0
google-generativeai>=0.3.0
